% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAAQD2.R
\name{PAAQD2}
\alias{PAAQD2}
\title{PAAQD2: a code-refined version of PAAQD}
\usage{
PAAQD2(peptideSet)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}
}
\description{
PAAQD2: a code-refined version of PAAQD
}
